Revert "chore: refactor logging tests (#229)"
This reverts commit 3984f1fa065fd3c7c94bd80e201b290cf2a86d9e.
