+++ pkgs/google_cloud/test/logging_test.dart
@@ -15,371 +15,615 @@
 @TestOn('vm')
 library;
 
-import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:google_cloud/general.dart';
 import 'package:google_cloud/http_serving.dart';
 import 'package:shelf/shelf.dart';
 import 'package:test/test.dart';
 
-const internalServerErrorMessage = 'Internal Server Error';
-
 void main() {
-  for (var environment in _Environment.values) {
-    for (var responseType in _ResponseType.values) {
-      for (var responseScenario in _ResponseScenarios.values) {
-        final name = [
-          responseScenario.name,
-          responseType.name,
-          environment.name,
-        ].join(' | ');
+  group('structuredLogEntry', () {
+    test('simple message', () {
+      final entry = structuredLogEntry('hello', LogSeverity.info);
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'message': 'hello', 'severity': 'INFO'});
+    });
 
-        test('[$name]', () async {
-          final middleware = createLoggingMiddleware(
-            projectId: environment.projectId,
-          );
+    test('message with traceId in payload', () {
+      final entry = structuredLogEntry(
+        'hello',
+        LogSeverity.info,
+        payload: {'logging.googleapis.com/trace': 'trace-123'},
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {
+        'message': 'hello',
+        'severity': 'INFO',
+        'logging.googleapis.com/trace': 'trace-123',
+      });
+    });
 
-          final handler = middleware(responseScenario.handler);
+    test('list message remains in message key', () {
+      final message = ['foo', 'bar'];
+      final entry = structuredLogEntry(message, LogSeverity.info);
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'message': message, 'severity': 'INFO'});
+    });
 
-          final logMatchers = _logMatcherFactory(responseScenario, environment);
+    test('map message is merged into payload', () {
+      final message = {'foo': 'bar', 'count': 42};
+      final entry = structuredLogEntry(message, LogSeverity.info);
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'foo': 'bar', 'count': 42, 'severity': 'INFO'});
+    });
 
-          final stdoutLines = <String>[];
-          final stderrLines = <String>[];
-          late Response response;
+    test('map message with message key extracts message', () {
+      final message = {'foo': 'bar', 'message': 'my msg'};
+      final entry = structuredLogEntry(message, LogSeverity.info);
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'foo': 'bar', 'message': 'my msg', 'severity': 'INFO'});
+    });
 
-          await IOOverrides.runZoned(
-            () => runZoned(
-              zoneSpecification: ZoneSpecification(
-                print: (_, _, _, String line) => stdoutLines.add(line),
-              ),
-              () async {
-                try {
-                  response = await handler(
-                    responseType.toRequest(
-                      responseScenario == _ResponseScenarios.successfulWithLogs,
-                    ),
-                  );
-                } catch (e, s) {
-                  if (responseScenario ==
-                          _ResponseScenarios.nonHttpResponseError &&
-                      environment == _Environment.normal) {
-                    stderrLines
-                      ..add('Exception: $e')
-                      ..add(s.toString());
-                    response = Response.internalServerError();
-                  } else {
-                    rethrow;
-                  }
-                }
-              },
-            ),
-            stdout: () => _MockStdout(stdoutLines),
-            stderr: () => _MockStdout(stderrLines),
-          );
+    test('payload overrides map message', () {
+      final message = {'foo': 'bar', 'count': 42, 'message': 'original'};
+      final entry = structuredLogEntry(
+        message,
+        LogSeverity.info,
+        payload: {'count': 99, 'env': 'prod', 'message': 'overridden'},
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {
+        'foo': 'bar',
+        'count': 99,
+        'env': 'prod',
+        'message': 'overridden',
+        'severity': 'INFO',
+      });
+    });
 
-          expect(
-            stdoutLines.join('\n'),
-            logMatchers.stdout,
-            reason: 'stdout matcher',
-          );
-          expect(
-            stderrLines.join('\n'),
-            logMatchers.stderr,
-            reason: 'stderr matcher',
-          );
+    test('non-encodable message is stringified', () {
+      final message = _NonEncodable();
+      final entry = structuredLogEntry(message, LogSeverity.info);
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'message': 'I am not encodable', 'severity': 'INFO'});
+    });
 
-          expect(
-            response,
-            _responseScenarioFactory(responseScenario, responseType),
-            reason: 'response matcher',
+    test('with payload', () {
+      final entry = structuredLogEntry(
+        'hello',
+        LogSeverity.info,
+        payload: {'foo': 'bar', 'count': 42},
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {
+        'foo': 'bar',
+        'count': 42,
+        'message': 'hello',
+        'severity': 'INFO',
+      });
+    });
+
+    test('with empty message', () {
+      final entry = structuredLogEntry(
+        '',
+        LogSeverity.info,
+        payload: {'foo': 'bar', 'count': 42},
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'foo': 'bar', 'count': 42, 'severity': 'INFO'});
+    });
+
+    test('payload does not override core fields', () {
+      final entry = structuredLogEntry(
+        'hello',
+        LogSeverity.info,
+        payload: {'message': 'overridden', 'severity': 'CRITICAL'},
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'message': 'hello', 'severity': 'INFO'});
+    });
+
+    test('non-encodable payload is stringified', () {
+      final payload = {'foo': _NonEncodable()};
+      final entry = structuredLogEntry(
+        'hello',
+        LogSeverity.info,
+        payload: payload,
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {
+        'foo': 'I am not encodable',
+        'message': 'hello',
+        'severity': 'INFO',
+      });
+    });
+
+    test('cyclic payload drops payload and stringifies message', () {
+      final payload = <String, dynamic>{};
+      payload['cycle'] = payload;
+      final message = _NonEncodable();
+      final entry = structuredLogEntry(
+        message,
+        LogSeverity.info,
+        payload: payload,
+      );
+      final map = jsonDecode(entry) as Map<String, dynamic>;
+      expect(map, {'message': 'I am not encodable', 'severity': 'INFO'});
+    });
+  });
+
+  group('LogSeverity', () {
+    test('toJson returns name', () {
+      expect(LogSeverity.info.toJson(), 'INFO');
+      expect(LogSeverity.error.toJson(), 'ERROR');
+    });
+
+    test('comparable', () {
+      expect(LogSeverity.info.compareTo(LogSeverity.error), isNegative);
+      expect(LogSeverity.critical.compareTo(LogSeverity.warning), isPositive);
+    });
+  });
+
+  group('CloudLogger.defaultLogger()', () {
+    const logger = CloudLogger.defaultLogger();
+
+    test('log with default severity', () {
+      expect(
+        () => logger.log('hello', LogSeverity.defaultSeverity),
+        prints('hello\n'),
+      );
+    });
+
+    test('log with explicit severity', () {
+      expect(
+        () => logger.log('hello', LogSeverity.error),
+        prints('ERROR: hello\n'),
+      );
+    });
+
+    test('log with payload', () {
+      expect(
+        () => logger.log('hello', LogSeverity.error, payload: {'foo': 'bar'}),
+        prints('ERROR: hello {foo: bar}\n'),
+      );
+    });
+  });
+
+  group('middleware', () {
+    test('cloudLoggingMiddleware logs structured entries', () async {
+      final handler = _testHandler((request) {
+        currentLogger.info('inner log');
+        return Response.ok('done');
+      });
+
+      await expectLater(
+        () => handler(
+          Request(
+            'GET',
+            Uri.parse('http://localhost/'),
+            headers: {
+              'x-cloud-trace-context':
+                  '0123456789abcdef0123456789abcdef/123;o=1',
+            },
+          ),
+        ),
+        prints(
+          isA<String>().having(
+            (output) => jsonDecode(output) as Map<String, dynamic>,
+            'as map',
+            {
+              'message': 'inner log',
+              'severity': 'INFO',
+              'logging.googleapis.com/trace':
+                  'projects/test-project/traces/0123456789abcdef0123456789abcdef',
+              'logging.googleapis.com/spanId': '000000000000007b',
+              'logging.googleapis.com/trace_sampled': true,
+            },
+          ),
+        ),
+      );
+    });
+
+    test(
+      'cloudLoggingMiddleware logs HttpResponseException with payload',
+      () async {
+        final handler = _testHandler((request) {
+          throw HttpResponseException(
+            403,
+            'Forbidden access',
+            status: 'PERMISSION_DENIED',
           );
         });
-      }
-    }
-  }
-}
 
-enum _ResponseScenarios {
-  httpResponseErrorMinimal(_throwHttpResponseMinimal),
-  httpResponseErrorWithDetails(_throwHttpResponseWithDetails),
-  nonHttpResponseError(_throwNonHttpResponseError),
-  successfulWithLogs(_respondSuccessfullyWithLogs);
-
-  final Handler handler;
-
-  const _ResponseScenarios(this.handler);
-}
-
-enum _Environment {
-  cloud('project-id'),
-  normal(null);
-
-  final String? projectId;
-
-  const _Environment(this.projectId);
-}
-
-enum _ResponseType {
-  json('application/json'),
-  text('text/plain');
-
-  final String contentType;
-
-  const _ResponseType(this.contentType);
-
-  Request toRequest(bool includeTraceContext) => Request(
-    'GET',
-    Uri.parse('http://localhost/'),
-    headers: {
-      'content-type': contentType,
-      if (includeTraceContext)
-        'x-cloud-trace-context': '0123456789abcdef0123456789abcdef/123;o=1',
-    },
-  );
-}
-
-TypeMatcher<Response> _responseScenarioFactory(
-  _ResponseScenarios responseScenario,
-  _ResponseType responseType,
-) => switch ((responseScenario, responseType)) {
-  (_ResponseScenarios.httpResponseErrorMinimal, _ResponseType.text) =>
-    _responseMatcher(
-      statusCode: 400,
-      contentType: anyOf(isNull, contains('text/plain')),
-      body: contains('minimal (400)'),
-    ),
-  (_ResponseScenarios.httpResponseErrorMinimal, _ResponseType.json) =>
-    _responseMatcher(
-      statusCode: 400,
-      contentType: contains('application/json'),
-      body: _jsonStringMatcher({
-        'error': {'code': 400, 'message': 'minimal'},
-      }),
-    ),
-  (_ResponseScenarios.httpResponseErrorWithDetails, _ResponseType.text) =>
-    _responseMatcher(
-      statusCode: 400,
-      contentType: anyOf(isNull, contains('text/plain')),
-      body: contains('with details (400)'),
-    ),
-  (_ResponseScenarios.httpResponseErrorWithDetails, _ResponseType.json) =>
-    _responseMatcher(
-      statusCode: 400,
-      contentType: contains('application/json'),
-      body: _jsonStringMatcher({
-        'error': {
-          'code': 400,
-          'message': 'with details',
-          'status': 'BAD_REQUEST',
-          'details': [
-            {
-              'type': 'type.googleapis.com/google.rpc.BadRequest',
-              'fieldViolations': <Never>[],
-            },
-          ],
-        },
-      }),
-    ),
-  (_ResponseScenarios.nonHttpResponseError, _ResponseType.text) =>
-    _responseMatcher(
-      statusCode: 500,
-      contentType: anyOf(isNull, contains('text/plain')),
-      body: contains(internalServerErrorMessage),
-    ),
-  (_ResponseScenarios.nonHttpResponseError, _ResponseType.json) =>
-    _responseMatcher(
-      statusCode: 500,
-      contentType: anyOf(isNull, contains('text/plain')),
-      body: contains(internalServerErrorMessage),
-    ),
-  (_ResponseScenarios.successfulWithLogs, _) => _responseMatcher(
-    statusCode: 200,
-    contentType: contains('text/plain'),
-    body: equals('done'),
-  ),
-};
-
-({Matcher stdout, Matcher stderr}) _logMatcherFactory(
-  _ResponseScenarios responseScenario,
-  _Environment environment,
-) => switch ((responseScenario, environment)) {
-  (_ResponseScenarios.httpResponseErrorMinimal, _Environment.cloud) => (
-    stdout: _jsonStringMatcher({
-      'message': 'minimal (400)',
-      'severity': 'WARNING',
-      'error': {'code': 400, 'message': 'minimal'},
-      'stack_trace': _stackTraceMatcher,
-      'logging.googleapis.com/sourceLocation': isA<Map<String, dynamic>>(),
-    }),
-    stderr: isEmpty,
-  ),
-  (_ResponseScenarios.httpResponseErrorWithDetails, _Environment.cloud) => (
-    stdout: _jsonStringMatcher({
-      'message': startsWith('with details (400)'),
-      'severity': 'WARNING',
-      'error': {
-        'code': 400,
-        'message': 'with details',
-        'status': 'BAD_REQUEST',
-        'details': [
-          {
-            'type': 'type.googleapis.com/google.rpc.BadRequest',
-            'fieldViolations': <Never>[],
-          },
-        ],
+        await expectLater(
+          () => handler(Request('GET', Uri.parse('http://localhost/'))),
+          prints(
+            isA<String>().having(
+              (output) => jsonDecode(output) as Map<String, dynamic>,
+              'as map',
+              {
+                'message': 'Forbidden access (403) [PERMISSION_DENIED]',
+                'severity': 'WARNING',
+                'error': {
+                  'code': 403,
+                  'message': 'Forbidden access',
+                  'status': 'PERMISSION_DENIED',
+                },
+                'stack_trace': isA<String>(),
+                'logging.googleapis.com/sourceLocation':
+                    isA<Map<String, dynamic>>(),
+              },
+            ),
+          ),
+        );
       },
-      'stack_trace': _stackTraceMatcher,
-      'logging.googleapis.com/sourceLocation': isA<Map<String, dynamic>>(),
-    }),
-    stderr: isEmpty,
-  ),
-  (_ResponseScenarios.nonHttpResponseError, _Environment.cloud) => (
-    stdout: _jsonStringMatcher({
-      'message': 'Exception: non http error',
-      'severity': 'ERROR',
-      'stack_trace': _stackTraceMatcher,
-      'logging.googleapis.com/sourceLocation': isA<Map<String, dynamic>>(),
-    }),
-    stderr: isEmpty,
-  ),
-  (_ResponseScenarios.httpResponseErrorMinimal, _Environment.normal) => (
-    stdout: endsWith('[400] /'),
-    stderr: allOf([contains('minimal (400)'), _stackTraceMatcher]),
-  ),
-  (_ResponseScenarios.httpResponseErrorWithDetails, _Environment.normal) => (
-    stdout: endsWith('[400] /'),
-    stderr: allOf([
-      contains('Invalid argument(s): inner error (ArgumentError)'),
-      contains('[BAD_REQUEST]'),
-      _stackTraceMatcher,
-    ]),
-  ),
-  (_ResponseScenarios.nonHttpResponseError, _Environment.normal) => (
-    stdout: contains('[ERROR]'),
-    stderr: allOf([contains('Exception: non http error'), _stackTraceMatcher]),
-  ),
-  (_ResponseScenarios.successfulWithLogs, _Environment.cloud) => (
-    stdout: isA<String>().having(
-      (s) => s
-          .trim()
-          .split('\n')
-          .map((l) => jsonDecode(l) as Map<String, dynamic>)
-          .toList(),
-      'parsed JSON lines',
-      [
-        {
-          'message': 'trace me',
-          'severity': 'INFO',
-          'logging.googleapis.com/trace':
-              'projects/project-id/traces/0123456789abcdef0123456789abcdef',
-          'logging.googleapis.com/spanId': '000000000000007b',
-          'logging.googleapis.com/trace_sampled': true,
-        },
-        {
-          'message': 'print me',
-          'severity': 'INFO',
-          'logging.googleapis.com/trace':
-              'projects/project-id/traces/0123456789abcdef0123456789abcdef',
-          'logging.googleapis.com/spanId': '000000000000007b',
-          'logging.googleapis.com/trace_sampled': true,
-        },
-        {
-          'message': 'default me',
-          'severity': 'DEFAULT',
-          'logging.googleapis.com/trace':
-              'projects/project-id/traces/0123456789abcdef0123456789abcdef',
-          'logging.googleapis.com/spanId': '000000000000007b',
-          'logging.googleapis.com/trace_sampled': true,
-        },
-        {
-          'message': 'warning me',
-          'severity': 'WARNING',
-          'logging.googleapis.com/trace':
-              'projects/project-id/traces/0123456789abcdef0123456789abcdef',
-          'logging.googleapis.com/spanId': '000000000000007b',
-          'logging.googleapis.com/trace_sampled': true,
-        },
-      ],
-    ),
-    stderr: isEmpty,
-  ),
-  (_ResponseScenarios.successfulWithLogs, _Environment.normal) => (
-    stdout: allOf([
-      contains('trace me'),
-      contains('print me'),
-      contains('default me'),
-      contains('WARNING: warning me'),
-      endsWith('[200] /'),
-    ]),
-    stderr: isEmpty,
-  ),
-};
-
-TypeMatcher<Response> _responseMatcher({
-  required int statusCode,
-  required Matcher contentType,
-  required Matcher body,
-}) => isA<Response>()
-    .having((r) => r.statusCode, 'statusCode', statusCode)
-    .having((r) => r.headers['content-type'], 'contentType', contentType)
-    .having((r) => r.readAsString(), 'body', completion(body));
-
-Matcher _jsonStringMatcher(Object expected) => isA<String>().having(
-  (output) => jsonDecode(output) as Map<String, dynamic>,
-  'decoded as a JSON map',
-  equals(expected),
-);
-
-final _stackTraceMatcher = matches(
-  RegExp(r'test/logging_test.dart[: ]\d+:\d+'),
-);
-
-Future<Response> _throwHttpResponseMinimal(_) =>
-    throw HttpResponseException(400, 'minimal');
-
-Future<Response> _throwHttpResponseWithDetails(_) {
-  try {
-    throw ArgumentError('inner error');
-  } catch (error, stack) {
-    throw HttpResponseException(
-      400,
-      'with details',
-      status: 'BAD_REQUEST',
-      details: [
-        {
-          'type': 'type.googleapis.com/google.rpc.BadRequest',
-          'fieldViolations': [],
-        },
-      ],
-      innerError: error,
-      innerStack: stack,
     );
-  }
+
+    test('cloudLoggingMiddleware logs notImplemented with payload', () async {
+      final handler = _testHandler((request) {
+        throw HttpResponseException.notImplemented(
+          message: 'Not implemented yet',
+        );
+      });
+
+      await expectLater(
+        () => handler(Request('GET', Uri.parse('http://localhost/'))),
+        prints(
+          isA<String>().having(
+            (output) => jsonDecode(output) as Map<String, dynamic>,
+            'as map',
+            {
+              'message': 'Not implemented yet (501) [UNIMPLEMENTED]',
+              'severity': 'ERROR',
+              'error': {
+                'code': 501,
+                'message': 'Not implemented yet',
+                'status': 'UNIMPLEMENTED',
+              },
+              'stack_trace': isA<String>(),
+              'logging.googleapis.com/sourceLocation':
+                  isA<Map<String, dynamic>>(),
+            },
+          ),
+        ),
+      );
+    });
+
+    test(
+      'httpResponseExceptionMiddleware handles HttpResponseException',
+      () async {
+        final handler = _testHandler((request) {
+          throw HttpResponseException(400, 'Custom bad request');
+        });
+
+        late Response response;
+        await expectLater(
+          () async => response = await handler(
+            Request('GET', Uri.parse('http://localhost/')),
+          ),
+          prints(
+            isA<String>().having(
+              (output) => jsonDecode(output) as Map<String, dynamic>,
+              'as map',
+              {
+                'message': 'Custom bad request (400)',
+                'severity': 'WARNING',
+                'error': {'code': 400, 'message': 'Custom bad request'},
+                'stack_trace': isA<String>(),
+                'logging.googleapis.com/sourceLocation':
+                    isA<Map<String, dynamic>>(),
+              },
+            ),
+          ),
+        );
+        expect(response.statusCode, 400);
+        expect(
+          await response.readAsString(),
+          contains('Custom bad request (400)'), // toString() output
+        );
+      },
+    );
+
+    test(
+      'httpResponseExceptionMiddleware handles HttpResponseException (JSON)',
+      () async {
+        final handler = _testHandler((request) {
+          throw HttpResponseException(400, 'Custom bad request');
+        });
+
+        late Response response;
+        await expectLater(
+          () async => response = await handler(
+            Request(
+              'GET',
+              Uri.parse('http://localhost/'),
+              headers: {'Accept': 'application/json'},
+            ),
+          ),
+          prints(
+            isA<String>().having(
+              (output) => jsonDecode(output) as Map<String, dynamic>,
+              'as map',
+              {
+                'message': 'Custom bad request (400)',
+                'severity': 'WARNING',
+                'error': {'code': 400, 'message': 'Custom bad request'},
+                'stack_trace': isA<String>(),
+                'logging.googleapis.com/sourceLocation':
+                    isA<Map<String, dynamic>>(),
+              },
+            ),
+          ),
+        );
+        expect(response.statusCode, 400);
+        expect(
+          response.headers[HttpHeaders.contentTypeHeader],
+          contains('application/json'),
+        );
+
+        final body = await response.readAsString();
+        final json = jsonDecode(body) as Map<String, dynamic>;
+        expect(json, {
+          'error': {'code': 400, 'message': 'Custom bad request'},
+        });
+      },
+    );
+
+    test('skips empty details in toJson', () {
+      final e1 = HttpResponseException(400, 'Custom bad request');
+      expect(e1.toJson()['error'], {
+        'code': 400,
+        'message': 'Custom bad request',
+      });
+
+      final e2 = HttpResponseException(400, 'Custom bad request', details: []);
+      expect(e2.toJson()['error'], {
+        'code': 400,
+        'message': 'Custom bad request',
+      });
+    });
+
+    test(
+      'httpResponseExceptionMiddleware handles expanded HttpResponseException',
+      () async {
+        final handler = _testHandler((request) {
+          throw HttpResponseException.badRequest(
+            message: 'Custom bad request',
+            status: 'INVALID_ARGUMENT',
+            details: [
+              {'field': 'name', 'message': 'required'},
+            ],
+          );
+        });
+
+        late Response response;
+        await expectLater(
+          () async => response = await handler(
+            Request(
+              'GET',
+              Uri.parse('http://localhost/'),
+              headers: {'Accept': 'application/json'},
+            ),
+          ),
+          prints(
+            isA<String>().having(
+              (output) => jsonDecode(output) as Map<String, dynamic>,
+              'as map',
+              {
+                'message':
+                    'Custom bad request (400) [INVALID_ARGUMENT] '
+                    'Details: [{field: name, message: required}]',
+                'severity': 'WARNING',
+                'error': {
+                  'code': 400,
+                  'message': 'Custom bad request',
+                  'status': 'INVALID_ARGUMENT',
+                  'details': [
+                    {'field': 'name', 'message': 'required'},
+                  ],
+                },
+                'stack_trace': isA<String>(),
+                'logging.googleapis.com/sourceLocation':
+                    isA<Map<String, dynamic>>(),
+              },
+            ),
+          ),
+        );
+        expect(response.statusCode, 400);
+
+        final body = await response.readAsString();
+        final json = jsonDecode(body) as Map<String, dynamic>;
+        expect(json, {
+          'error': {
+            'code': 400,
+            'message': 'Custom bad request',
+            'status': 'INVALID_ARGUMENT',
+            'details': [
+              {'field': 'name', 'message': 'required'},
+            ],
+          },
+        });
+      },
+    );
+
+    test('httpResponseExceptionMiddleware logs to stderr', () async {
+      final stderrLines = <String>[];
+      final handler = const Pipeline()
+          .addMiddleware(createLoggingMiddleware())
+          .addHandler((request) {
+            throw HttpResponseException(400, 'Custom bad request');
+          });
+
+      await expectLater(
+        () => IOOverrides.runZoned(() async {
+          final response = await handler(
+            Request('GET', Uri.parse('http://localhost/')),
+          );
+          expect(response.statusCode, 400);
+        }, stderr: () => _MockStdout(stderrLines)),
+        prints(contains('[400] /')),
+      );
+
+      expect(stderrLines, hasLength(1));
+      final lines = stderrLines.single.split('\n');
+      expect(lines.first, contains('Custom bad request (400)'));
+      expect(lines[1], contains('logging_test.dart'));
+    });
+  });
+
+  group('HttpResponseException', () {
+    test('valid status code', () {
+      final ex = HttpResponseException(400, 'Bad');
+      expect(ex.statusCode, 400);
+      expect(ex.message, 'Bad');
+      expect(ex.toString(), 'Bad (400)');
+    });
+
+    test('invalid status code low', () {
+      expect(() => HttpResponseException(399, 'Bad'), throwsArgumentError);
+    });
+
+    test('invalid status code high', () {
+      expect(() => HttpResponseException(600, 'Bad'), throwsArgumentError);
+    });
+
+    test('empty message', () {
+      expect(
+        () => HttpResponseException(400, ''),
+        throwsA(isA<AssertionError>()),
+      );
+    });
+  });
+
+  group('parseTraceContext', () {
+    test('parses full context', () {
+      final context = TraceContextData.parse(
+        projectId: 'test-project',
+        traceHeader: '0123456789abcdef0123456789abcdef/1054454457908058113;o=1',
+      );
+      expect(
+        context.traceId,
+        'projects/test-project/traces/0123456789abcdef0123456789abcdef',
+      );
+      expect(context.spanId, '0ea22cbe236fd801');
+      expect(context.traceSampled, true);
+    });
+
+    test('parses without sampled flag', () {
+      final context = TraceContextData.parse(
+        projectId: 'test-project',
+        traceHeader: '0123456789abcdef0123456789abcdef/123',
+      );
+      expect(
+        context.traceId,
+        'projects/test-project/traces/0123456789abcdef0123456789abcdef',
+      );
+      expect(context.spanId, '000000000000007b');
+      expect(context.traceSampled, isFalse);
+    });
+
+    test(
+      'parses format without trace options but trailing semicolon flag off',
+      () {
+        final context = TraceContextData.parse(
+          projectId: 'test-project',
+          traceHeader: '0123456789abcdef0123456789abcdef/123;o=0',
+        );
+        expect(
+          context.traceId,
+          'projects/test-project/traces/0123456789abcdef0123456789abcdef',
+        );
+        expect(context.spanId, '000000000000007b');
+        expect(context.traceSampled, false);
+      },
+    );
+
+    test('parses minimal trace', () {
+      final context = TraceContextData.parse(
+        projectId: 'test-project',
+        traceHeader: '0123456789abcdef0123456789abcdef',
+      );
+      expect(
+        context.traceId,
+        'projects/test-project/traces/0123456789abcdef0123456789abcdef',
+      );
+      expect(context.spanId, isNull);
+      expect(context.traceSampled, isFalse);
+    });
+  });
+
+  group('TraceContextData.asPayloadMap', () {
+    test('full context includes everything', () {
+      final context = TraceContextData(
+        traceId: 'test-trace',
+        spanId: 'test-span',
+        traceSampled: true,
+      );
+      expect(context.asPayloadMap(), {
+        'logging.googleapis.com/trace': 'test-trace',
+        'logging.googleapis.com/spanId': 'test-span',
+        'logging.googleapis.com/trace_sampled': true,
+      });
+    });
+
+    test('omits spanId when null', () {
+      final context = TraceContextData(traceId: 'test-trace');
+      expect(context.asPayloadMap(), {
+        'logging.googleapis.com/trace': 'test-trace',
+      });
+    });
+
+    test('omits traceSampled when false', () {
+      final context = TraceContextData(
+        traceId: 'test-trace',
+        spanId: 'test-span',
+        // traceSampled: false, // Default
+      );
+      expect(context.asPayloadMap(), {
+        'logging.googleapis.com/trace': 'test-trace',
+        'logging.googleapis.com/spanId': 'test-span',
+      });
+    });
+
+    test('merges with existing payload', () {
+      final context = TraceContextData(
+        traceId: 'test-trace',
+        spanId: 'test-span',
+      );
+      final payload = {'message': 'hello', 'count': 42};
+      expect(context.asPayloadMap(payload), {
+        'message': 'hello',
+        'count': 42,
+        'logging.googleapis.com/trace': 'test-trace',
+        'logging.googleapis.com/spanId': 'test-span',
+      });
+    });
+  });
 }
 
-Future<Response> _throwNonHttpResponseError(_) =>
-    throw Exception('non http error');
+Handler _testHandler(Handler inner) => const Pipeline()
+    .addMiddleware(cloudLoggingMiddleware('test-project'))
+    .addHandler(inner);
 
-Future<Response> _respondSuccessfullyWithLogs(_) async {
-  currentLogger.info('trace me');
-  print('print me');
-  currentLogger
-    ..log('default me', LogSeverity.defaultSeverity)
-    ..warning('warning me');
-  return Response.ok('done', headers: {'content-type': 'text/plain'});
+class _NonEncodable {
+  @override
+  String toString() => 'I am not encodable';
 }
 
 class _MockStdout implements Stdout {
-  _MockStdout(this._lines);
-
   final List<String> _lines;
+
+  _MockStdout(this._lines);
 
   @override
   bool get supportsAnsiEscapes => false;
 
   @override
-  void writeln([Object? object = '']) => _lines.add('$object');
+  void writeln([Object? object = '']) {
+    _lines.add('$object');
+  }
 
   @override
   dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
